chore: init graphql

+++ api/entity/password_reset.go
@@ -3,5 +3,5 @@ package entity
 type PasswordReset struct {
 	Email   string `json:"email"`
 	Token   string `json:"token"`
-	Expired int32  `json:"expired"`
+	Expired int64  `json:"expired"`
 }
